+++ lib/core/admob/services/admob_banner.dart
@@ -1,0 +1,78 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:food_gram_app/core/admob/config/admob_config.dart';
+import 'package:food_gram_app/core/data/purchase/subscription_provider.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart';
+
+/// バナー広告の状態を管理するプロバイダー
+final bannerAdProvider =
+    StateNotifierProvider<BannerAdNotifier, BannerAd?>((ref) {
+  return BannerAdNotifier();
+});
+
+/// バナー広告の状態を管理するNotifier
+class BannerAdNotifier extends StateNotifier<BannerAd?> {
+  BannerAdNotifier() : super(null) {
+    _loadAd();
+  }
+
+  void _loadAd() {
+    BannerAd(
+      adUnitId: AdmobConfig.bannerAdUnitId,
+      size: AdSize.banner,
+      request: const AdRequest(),
+      listener: BannerAdListener(
+        onAdFailedToLoad: (ad, error) {
+          ad.dispose();
+          state = null;
+        },
+        onAdLoaded: (ad) {
+          state = ad as BannerAd;
+        },
+      ),
+    ).load();
+  }
+
+  @override
+  void dispose() {
+    state?.dispose();
+    super.dispose();
+  }
+}
+
+/// バナー広告を表示するウィジェット
+class AdmobBanner extends ConsumerWidget {
+  const AdmobBanner({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final bannerAd = ref.watch(bannerAdProvider);
+    final subscriptionState = ref.watch(subscriptionProvider);
+
+    return subscriptionState.when(
+      data: (isSubscribed) => _buildBannerContainer(bannerAd, isSubscribed),
+      error: (_, __) => const SizedBox.shrink(),
+      loading: () => _buildLoadingContainer(bannerAd),
+    );
+  }
+
+  Widget _buildBannerContainer(BannerAd? bannerAd, bool isSubscribed) {
+    if (isSubscribed || bannerAd == null) {
+      return const SizedBox.shrink();
+    }
+
+    return Container(
+      width: double.infinity,
+      alignment: Alignment.center,
+      height: bannerAd.size.height.toDouble(),
+      child: AdWidget(ad: bannerAd),
+    );
+  }
+
+  Widget _buildLoadingContainer(BannerAd? bannerAd) {
+    return SizedBox(
+      width: double.infinity,
+      height: bannerAd?.size.height.toDouble() ?? 50.0,
+    );
+  }
+}
